# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/v1/admin/addApiGroup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "创建api",
                "parameters": [
                    {
                        "description": "api数组",
                        "name": "{array}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.FormAPIGroupAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/addRoleApiPerm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "添加角色接口权限",
                "parameters": [
                    {
                        "description": "添加api权限表单",
                        "name": "{}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.FormRoleAPIPerm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/addRoleGroup": {
            "post": {
                "description": "添加角色组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "添加角色组",
                "parameters": [
                    {
                        "description": "添加角色组",
                        "name": "{}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.FormRoleAPIPerm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/delMenu": {
            "delete": {
                "description": "删除菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/editApiGroup": {
            "put": {
                "description": "修改api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "修改api",
                "parameters": [
                    {
                        "description": "表单",
                        "name": "{}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.FormAPIGroupEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/editMenu": {
            "put": {
                "description": "更新菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "description": "菜单form",
                        "name": "{}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.FormUpdateMenu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/getAllApiGroup": {
            "get": {
                "description": "获取全部api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "获取全部api",
                "parameters": [
                    {
                        "description": "获取api列表",
                        "name": "{}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.ApiListForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/removeApiGroup": {
            "delete": {
                "description": "删除api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "删除api",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/removeRoleAPIPerm": {
            "delete": {
                "description": "删除角色接口权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "删除角色接口权限",
                "parameters": [
                    {
                        "description": "删除api权限表单",
                        "name": "{}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.FormRoleAPIPerm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/getRoleApiList": {
            "get": {
                "description": "获取用户权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "获取用户权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户uid,不输入默认返回自己uid",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Meta": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "app.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "meta": {
                    "type": "object",
                    "$ref": "#/definitions/app.Meta"
                }
            }
        },
        "body.Body": {
            "type": "object",
            "additionalProperties": true
        },
        "utils.ApiListForm": {
            "type": "object",
            "properties": {
                "apiGroup": {
                    "description": "组名",
                    "type": "string"
                },
                "description": {
                    "description": "中文描述",
                    "type": "string"
                },
                "method": {
                    "description": "请求方式",
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "path": {
                    "description": "api路径",
                    "type": "string"
                }
            }
        },
        "utils.FormAPIGroupAdd": {
            "type": "object",
            "required": [
                "apiGroup",
                "description",
                "method",
                "path"
            ],
            "properties": {
                "apiGroup": {
                    "description": "组名",
                    "type": "string"
                },
                "description": {
                    "description": "中文描述",
                    "type": "string"
                },
                "method": {
                    "description": "请求方式",
                    "type": "string"
                },
                "path": {
                    "description": "api路径",
                    "type": "string"
                }
            }
        },
        "utils.FormAPIGroupEdit": {
            "type": "object",
            "required": [
                "apiGroup",
                "description",
                "id",
                "method",
                "path"
            ],
            "properties": {
                "apiGroup": {
                    "description": "组名",
                    "type": "string"
                },
                "description": {
                    "description": "中文描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "description": "请求方式",
                    "type": "string"
                },
                "path": {
                    "description": "api路径",
                    "type": "string"
                }
            }
        },
...
        "utils.FormRoleAPIPerm": {
            "type": "object",
            "required": [
                "groupIds",
                "role"
            ],
            "properties": {
                "groupIds": {
                    "description": "api id",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role": {
                    "description": "角色名称",
                    "type": "string"
                }
            }
        },
        "utils.FormUpdateMenu": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "component": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meta": {
                    "type": "object",
                    "$ref": "#/definitions/body.Body"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
    }
}